Add stripe

+++ StoreProject.Application/DTOs/Coupon/CouponDto.cs
@@ -10,7 +10,7 @@ namespace StoreProject.Application.DTOs.Coupon
     public class CouponDto : BaseDto
     {
         public string CouponCode { get; set; }
-        public double MinAmount { get; set; }
-        public double DiscountAmount { get; set; }
+        public decimal MinAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
     }
 }
